refactor: move ranking compute from CachedSeasonLeaderboardsRepository to Service
## What changed
- Moved getSortValue() and computeQaPerGame() from CachedSeasonLeaderboardsRepository
  to SeasonLeaderboardsService as private methods
- Added getFilteredLeaderboard() to SeasonLeaderboardsServiceInterface — this method
  owns filtering, sorting, and limiting over the raw repository result set
- CachedSeasonLeaderboardsRepository is now a pure cache pass-through (cache check →
  delegate → store), with no business logic
- SeasonLeaderboardsService takes a Repository dependency via constructor injection
- Module index.php wiring updated: calls service->getFilteredLeaderboard() instead of
  repository->getSeasonLeaders()

## Test migration
- Filter/sort/limit tests moved from CachedSeasonLeaderboardsRepositoryTest to
  SeasonLeaderboardsServiceTest (12 tests migrated)
- CachedSeasonLeaderboardsRepositoryTest now tests only cache behavior (7 tests)
- SeasonLeaderboardsViewTest updated for new Service constructor
- All 5315 tests pass
