+++ ibl5/classes/SeasonLeaderboards/CachedSeasonLeaderboardsRepository.php
@@ -10,10 +10,8 @@ use SeasonLeaderboards\Contracts\SeasonLeaderboardsRepositoryInterface;
 /**
  * CachedSeasonLeaderboardsRepository - Caching decorator for SeasonLeaderboardsRepositoryInterface.
  *
- * Caches the full unsorted leaders result set, years list, and teams list.
- * On cache hit, filters by year/team, sorts by the requested stat, and
- * slices for limit — all in PHP. This avoids materializing the expensive
- * ibl_hist VIEW on every sort/filter change.
+ * Pure cache pass-through: caches the full result set from the inner repository.
+ * Filtering, sorting, and limiting are handled by SeasonLeaderboardsService.
  *
  * @phpstan-import-type HistRow from SeasonLeaderboardsRepositoryInterface
  * @phpstan-import-type LeaderboardFilters from SeasonLeaderboardsRepositoryInterface
@@ -51,42 +49,6 @@ class CachedSeasonLeaderboardsRepository implements SeasonLeaderboardsRepository
             $innerResult = $this->inner->getSeasonLeaders([], 0);
             $rows = $innerResult['results'];
             $this->cache->set(self::CACHE_KEY_LEADERS, $rows, self::TTL_SECONDS);
-        }
-
-        // Filter by year
-        $yearFilter = (string) ($filters['year'] ?? '');
-        if ($yearFilter !== '') {
-            $yearInt = (int) $yearFilter;
-            $rows = array_values(array_filter(
-                $rows,
-                static fn (array $row): bool => $row['year'] === $yearInt
-            ));
-        }
-
-        // Filter by team
-        $teamId = (int) ($filters['team'] ?? 0);
-        if ($teamId !== 0) {
-            $rows = array_values(array_filter(
-                $rows,
-                static fn (array $row): bool => $row['teamid'] === $teamId
-            ));
-        }
-
-        // Sort by the requested stat DESC
-        $sortBy = (string) ($filters['sortby'] ?? 'PPG');
-        usort($rows, static function (array $a, array $b) use ($sortBy): int {
-            $aVal = self::getSortValue($a, $sortBy);
-            $bVal = self::getSortValue($b, $sortBy);
-            $cmp = $bVal <=> $aVal;
-            if ($cmp !== 0) {
-                return $cmp;
-            }
-            return $a['pid'] <=> $b['pid'];
-        });
-
-        // Apply limit
-        if ($limit > 0) {
-            $rows = array_slice($rows, 0, $limit);
         }
 
         return [
@@ -153,72 +115,5 @@ class CachedSeasonLeaderboardsRepository implements SeasonLeaderboardsRepository
         $this->cache->delete(self::CACHE_KEY_LEADERS);
         $this->cache->delete(self::CACHE_KEY_YEARS);
         $this->cache->delete(self::CACHE_KEY_TEAMS);
-    }
-
-    /**
-     * Compute the sort value for a row, matching the SQL expressions in
-     * SeasonLeaderboardsRepository::getSortColumn().
-     *
-     * @param HistRow $row
-     */
-    private static function getSortValue(array $row, string $sortBy): float
-    {
-        $games = $row['games'];
-        if ($games === 0 && $sortBy !== 'GAMES') {
-            return 0.0;
-        }
-
-        $fgm = $row['fgm'];
-        $fga = $row['fga'];
-        $ftm = $row['ftm'];
-        $fta = $row['fta'];
-        $tgm = $row['tgm'];
-        $tga = $row['tga'];
-
-        return match ($sortBy) {
-            'PPG' => (2 * $fgm + $ftm + $tgm) / $games,
-            'REB' => $row['reb'] / $games,
-            'OREB' => $row['orb'] / $games,
-            'DREB' => ($row['reb'] - $row['orb']) / $games,
-            'AST' => $row['ast'] / $games,
-            'STL' => $row['stl'] / $games,
-            'BLK' => $row['blk'] / $games,
-            'TO' => $row['tvr'] / $games,
-            'FOUL' => $row['pf'] / $games,
-            'QA' => self::computeQaPerGame($row, $games),
-            'FGM' => $fgm / $games,
-            'FGA' => $fga / $games,
-            'FGP' => $fga > 0 ? $fgm / $fga : 0.0,
-            'FTM' => $ftm / $games,
-            'FTA' => $fta / $games,
-            'FTP' => $fta > 0 ? $ftm / $fta : 0.0,
-            'TGM' => $tgm / $games,
-            'TGA' => $tga / $games,
-            'TGP' => $tga > 0 ? $tgm / $tga : 0.0,
-            'GAMES' => (float) $games,
-            'MIN' => $row['minutes'] / $games,
-            default => (2 * $fgm + $ftm + $tgm) / $games,
-        };
-    }
-
-    /**
-     * QA per game — matches SQL expression:
-     * ((2*fgm+ftm+tgm)+reb+(2*ast)+(2*stl)+(2*blk))-((fga-fgm)+(fta-ftm)+tvr+pf)) / games
-     *
-     * @param HistRow $row
-     */
-    private static function computeQaPerGame(array $row, int $games): float
-    {
-        $fgm = $row['fgm'];
-        $fga = $row['fga'];
-        $ftm = $row['ftm'];
-        $fta = $row['fta'];
-        $tgm = $row['tgm'];
-
-        $pts = 2 * $fgm + $ftm + $tgm;
-        $positive = $pts + $row['reb'] + (2 * $row['ast']) + (2 * $row['stl']) + (2 * $row['blk']);
-        $negative = ($fga - $fgm) + ($fta - $ftm) + $row['tvr'] + $row['pf'];
-
-        return ($positive - $negative) / $games;
     }
 }
